Fix handle delete poll

+++ app/core/services/deletePoll.js
@@ -1,5 +1,9 @@
+const PollOwnerException = require('../errors/PollOwnerException');
+
 function createDeletePoll(pollRepository) {
   return async function deletePoll(poll) {
+    const [pollToDelete] = await pollRepository.find({ id: poll.id });
+    if (pollToDelete.owner !== poll.owner) throw new PollOwnerException();
     return pollRepository.destroy(poll);
   };
 }
